move process_blob back to pending_blocks, process the block if possible
