+++ lib/lambda_ethereum_consensus/store/blobs.ex
@@ -4,34 +4,16 @@ defmodule LambdaEthereumConsensus.Store.Blobs do
   """
   require Logger
 
-  alias LambdaEthereumConsensus.P2P.BlobDownloader
   alias LambdaEthereumConsensus.Store.BlobDb
-  alias LambdaEthereumConsensus.Store.Blocks
   alias Types.BlockInfo
 
-  def process_blobs(store, {:ok, blobs}), do: {:ok, add_blobs(store, blobs)}
-
-  def process_blobs(store, {:error, reason}) do
-    # We might want to declare a block invalid here.
-    Logger.error("[PendingBlocks] Error downloading blobs: #{inspect(reason)}")
-    {:ok, store}
-  end
-
-  def add_blob(store, blob), do: add_blobs(store, [blob])
+  def add_blob(blob), do: add_blobs([blob])
 
   # To be used when a series of blobs are downloaded. Stores each blob.
-  def add_blobs(store, blobs) do
+  def add_blobs(blobs) do
     blobs
     |> Enum.map(&BlobDb.store_blob/1)
     |> Enum.uniq()
-    |> Enum.each(fn root ->
-      with %BlockInfo{status: :download_blobs} = block_info <- Blocks.get_block_info(root),
-           [] <- missing_for_block(block_info) do
-        block_info |> Blocks.change_status(:pending) |> Blocks.new_block_info()
-      end
-    end)
-
-    store
   end
 
   @spec missing_for_block(BlockInfo.t()) :: [Types.BlobIdentifier.t()]
@@ -50,15 +32,5 @@ defmodule LambdaEthereumConsensus.Store.Blobs do
       _ ->
         true
     end
-  end
-
-  def schedule_blob_download(missing_blobs, retries) do
-    BlobDownloader.request_blobs_by_root(
-      missing_blobs,
-      &process_blobs/2,
-      retries
-    )
-
-    :ok
   end
 end
